Hierarchical Posterior Model Fitting API

Data IO

HierarchicalPosteriorModel.get_Ps — Function

get_Ps(fit_results, matches, θh_pos_is_ventral; idx_use=[1,2,3,4,7], datasets_use=nothing, rngs_use=nothing, max_rng=2)

Return a vector of matrices containing the parameter sets for given matches and conditions.

Arguments

fit_results: A dictionary containing the CePNEM fit results for each dataset.

matches: A list of tuples corresponding to a given neuron, with dataset and the corresponding index in the traces.

θh_pos_is_ventral: A dictionary mapping dataset to a boolean indicating if positive θh corresponds to ventral head curvature (as opposed to dorsal).

idx_use: An optional array of indices of the parameters to be used. Default is [1,2,3,4,7].

datasets_use: An optional list of dataset indices to be used. If not provided, all datasets in matches are used.

rngs_use: An optional dictionary mapping dataset indices to lists of range indices to be used. If not provided, all ranges are used.

max_rng: An optional integer specifying the maximum number of ranges to be used. Default is 2.

Returns

Ps: A vector of matrices containing the parameter sets for the given matches and conditions.

HierarchicalPosteriorModel.get_datasets — Function

get_datasets(matches; datasets_use=nothing, rngs_use=nothing)

Retrieve datasets based on the given matches and optional filtering criteria.

Arguments

fit_results::Dict: A dictionary containing the CePNEM fit results for each dataset.

matches::Vector: A list of tuples containing the dataset identifier and the number of samples.

datasets_use::Union{Vector{String}, Nothing}: An optional list of dataset identifiers to include. If not provided, all datasets in matches will be used.

rngs_use::Union{Dict, Nothing}: An optional dictionary of dataset identifiers with a list of range indices to include. If not provided, all ranges will be used.

Returns

datasets: A list of tuples containing the selected dataset identifier, range index, and number of samples.

HierarchicalPosteriorModel.convert_hbdatasets_to_behs — Function

convert_hbdatasets_to_behs(fit_results::Dict, hbdatasets::Vector; max_len::Int=800)

Convert the given hbdatasets into separate v, θh, and P arrays based on fit_results.

Arguments

fit_results::Dict: A dictionary containing the results of a model fitting process.

hbdatasets::Vector: A list of tuples containing the selected dataset identifier, range index, and number of samples.

θh_pos_is_ventral::Dict{String, Bool}: Whether the θh values in the datasets are in ventral or dorsal coordinates.

max_len::Int (optional, default 800): The maximum length of each dataset.

Returns

v: A vector containing the concatenated v values from the datasets.

θh: A vector containing the concatenated θh values from the datasets.

P: A vector containing the concatenated P values from the datasets.

Model Fitting

HierarchicalPosteriorModel.initialize_params — Function

initialize_params(Ps::Vector{Matrix{Float64}}; idx_scaling::Vector{Int64}=[2,3,4])

Initialize the model parameters based on the given parameter samples Ps.

Arguments

Ps::Vector{Matrix{Float64}}: A vector of matrices containing the CePNEM parameter samples for each dataset.

idx_scaling::Vector{Int64}: An optional vector of indices indicating the parameters that need to be transformed to Cartesian coordinates for comparison into mvns. Currently, it is not supported for this to be any value other than its default [2,3,4].

Returns

HBParams: A HBParams instance containing the initialized values for mu, sigma, and x parameters.

HierarchicalPosteriorModel.fit_multivariate_normals — Function

fit_multivariate_normals(Ps::Vector{Matrix{Float64}})

Fits multivariate normal distributions to a set of input matrices.

Arguments

Ps::Vector{Matrix{Float64}}: A vector of matrices, each representing a set of data points.

Returns

A vector of fitted MvNormal objects.

HierarchicalPosteriorModel.optimize_MAP — Function

optimize_MAP(Ps::Vector{Matrix{Float64}}, params_init::HBParams; idx_scaling::Vector{Int64}=[2,3,4])

Optimize the maximum a posteriori (MAP) estimate for the given parameter samples Ps and initial parameters params_init.

Arguments

Ps::Vector{Matrix{Float64}}: A vector of matrices containing the CePNEM parameter samples for each dataset.

params_init::HBParams: A HBParams instance containing the initial values for mu, sigma, and x parameters.

idx_scaling::Vector{Int64}: An optional vector of indices indicating the parameters that need to be transformed to Cartesian coordinates for comparison into mvns. Currently, it is not supported for this to be any value other than its default [2,3,4].

max_iters::Int64: An optional integer indicating the maximum number of iterations for the optimization process. Default 200.

mvns::Union{Vector, Nothing}: An optional vector of multivariate normal distributions corresponding to the data. If nothing, they will be computed.

Returns

params_opt_struct: A HBParams instance containing the optimized values for mu, sigma, and x parameters.

result: The result of the optimization process.