Hierarchical Posterior Model API

HierarchicalPosteriorModel.HBParams — Type

struct HBParams

A structure for holding the mean (mu), standard deviation (sigma), and the spherical coordinates (x) of a model.

Fields

mu::Vector{Float64}: A vector of the global mean estimate for each parameter.

sigma::Vector{Float64}: A vector of the global standard deviation estimate for each parameter.

x::Vector{Vector{Float64}}: A vector of vectors representing the best parameters for each individual dataset.

HierarchicalPosteriorModel.joint_logprob — Function

joint_logprob(params::HBParams, data::Vector{Matrix{Float64}}, mvns::Vector; idx_scaling::Vector{Int64}=[2,3,4])

Compute the joint log probability of a HBParams instance given the data and multivariate normal distributions (mvns).

Arguments

params::HBParams: A HBParams instance containing the mu, sigma, and x parameters.

data::Vector{Matrix{Float64}}: A vector of matrices representing the data (CePNEM fit parameters) for each dataset.

mvns::Vector: A vector of multivariate normal distributions corresponding to the data.

idx_scaling::Vector{Int64}: An optional vector of indices indicating the parameters that need to be transformed to Cartesian coordinates for comparison into mvns.

Currently, it is not supported for this to be any value other than its default, [2,3,4].

Returns

logprob: The computed joint log probability of the given parameters.

HierarchicalPosteriorModel.joint_logprob_flat — Function

joint_logprob_flat(params_flat::Vector, data::Vector{Matrix{Float64}}, mvns::Vector, idx_scaling::Vector{Int64})

Compute the joint log probability of a flat parameter vector given the data, multivariate normal distributions (mvns), and scaling indices (idx_scaling).

Arguments

params_flat::Vector: A flat vector containing the concatenated mu, sigma, and x parameters.

data::Vector{Matrix{Float64}}: A vector of matrices representing the data (CePNEM fit parameters) for each dataset.

mvns::Vector: A vector of multivariate normal distributions corresponding to the data.

idx_scaling::Vector{Int64}: A vector of indices indicating the parameters that need to be transformed to Cartesian coordinates for comparison into mvns. Currently, it is not supported for this to be any value other than [2,3,4].

Returns

logprob: The computed joint log probability of the given parameters.

HierarchicalPosteriorModel.convert_hbparams_to_ps — Function

convert_hbparams_to_ps(hbparams::Vector{Float64})

Convert the given hbparams 5-vector of mu or x from HBParams into a new psvector with 8 elements compatible withmodel_nl8`.

Arguments

hbparams::Vector{Float64}: A vector of hierarchical model parameters.

Returns

ps: A new 8-element vector containing the converted hbparams.

HierarchicalPosteriorModel.compute_cartesian_average — Function

computecartesianaverage(hbparams::HBParams)

Compute the Cartesian average of the given hierarchical Bayesian parameters.

Arguments

hbparams::HBParams: A hierarchical Bayesian parameter object.

Returns

ps_tot::Vector{Float64}: The Cartesian average of the given hierarchical Bayesian parameters.