Variability API

HierarchicalPosteriorModel.get_variability — Function

get_variability(sigma::Array{Float64})

Compute the variability of the parameters based on the given sigma array from a hierarchical posterior model fit.

Arguments

sigma::Array{Float64}: An array containing the sigma values of the parameters.

Returns

variability: The computed variability of the parameters.

get_variability(model_params::HBParams)

Compute the variability metric based on the provided model parameters.

Arguments

model_params: A HBParams instance containing the standard deviation (sigma) parameter.

Returns

variability: A scalar value representing the computed variability metric.

HierarchicalPosteriorModel.get_variability_subtypes — Function

get_variability_subtypes(hierarchical_datasets::Vector, hierarchical_params::HBParams; neuron::String="")

Calculate the variability of parameters between datasets for different subtypes of variability: inter-dataset variability, intra-dataset variability, and left vs right (LR) variability.

Arguments

hierarchical_datasets::Vector: A vector containing the datasets used to fit the hierarchical model.

hierarchical_params::HBParams: An instance of HBParams containing the hierarchical model parameters.

neuron::String: (optional) The name of the neuron (default: empty string).

Returns

inter_variability: The calculated inter-dataset variability.

intra_variability: The calculated intra-dataset variability.

LR_variability: The calculated left-right variability.

inter_dataset_variability: A list of values from which inter-dataset variability is computed.

intra_dataset_variability: A list of values from which intra-dataset variability is computed.

LR_dataset_variability: A list of values from which left-right variability is computed.